+++ apps/bondy/src/bondy_rpc_load_balancer.erl
@@ -62,7 +62,7 @@
 -define(OPTS_SPEC, #{
     %% BONDY extension
     %% TODO this should be a map
-    %% #{strategy => #{id => queue_least_loaded, x_prefer_local}}
+    %% #{strategy => #{id => queue_least_loaded, _prefer_local}}
     strategy => #{
         required => true,
         allow_null => false,
@@ -89,14 +89,14 @@
                 true
         end
     },
-    x_prefer_local => #{
+    '_prefer_local' => #{
         required => true,
         allow_null => false,
         allow_undefined => false,
         default => true,
         datatype => boolean
     },
-    x_routing_key => #{
+    '_routing_key' => #{
         required => false,
         allow_null => false,
         allow_undefined => false,
@@ -122,8 +122,8 @@
                                 | queue_least_loaded_sample.
 -type opts()                ::  #{
     strategy := strategy(),
-    'x_prefer_local' => boolean(),
-    'x_routing_key' => binary()
+    '_prefer_local' => boolean(),
+    '_routing_key' => binary()
 }.
 -opaque iterator()          ::  #iterator{}.
 
@@ -206,12 +206,12 @@ iterate(#iterator{} = Iter) ->
 
 validate_options(Opts0) ->
     case maps_utils:validate(Opts0, ?OPTS_SPEC) of
-        #{strategy := jump_consistent_hash, 'x_routing_key' := _} = Opts ->
+        #{strategy := jump_consistent_hash, '_routing_key' := _} = Opts ->
             Opts;
         #{strategy := jump_consistent_hash} ->
             ErrorMap = bondy_error:map({
                 <<"missing_option">>,
-                <<"A value for option 'x_routing_key' or 'rkey' is required">>
+                <<"A value for option '_routing_key' or 'rkey' is required">>
             }),
             error(ErrorMap);
         Opts ->
@@ -240,7 +240,7 @@ prepare_entries(Entries, #{strategy := jump_consistent_hash}) ->
 prepare_entries(Entries, #{strategy := queue_least_loaded_sample}) ->
     lists_utils:shuffle(Entries);
 
-prepare_entries(Entries, #{x_prefer_local := Flag}) ->
+prepare_entries(Entries, #{'_prefer_local' := Flag}) ->
     maybe_sort_by_locality(Flag, Entries);
 
 prepare_entries(Entries, _) ->
@@ -371,7 +371,7 @@ next_consistent_hash(#iterator{entries = []}, _) ->
     '$end_of_table';
 
 next_consistent_hash(Iter, Algo) ->
-    Key = maps:get('x_routing_key', Iter#iterator.options),
+    Key = maps:get('_routing_key', Iter#iterator.options),
     Buckets = length(Iter#iterator.entries),
     Bucket = bondy_consistent_hashing:bucket(Key, Buckets, Algo),
     Entries = Iter#iterator.entries,